Fix draw visualiser title bar ignoring drags in some cases
