+++ osu.Framework/Graphics/Visualisation/TreeContainerStatus.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+namespace osu.Framework.Graphics.Visualisation
+{
+    internal enum TreeContainerStatus
+    {
+        Onscreen,
+        Offscreen
+    }
+}
